Memisahkan beberapa method ke class-class berbeda
