+++ src/lib/EmployeeInfo.java
@@ -1,0 +1,23 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package lib;
+
+/**
+ *
+ * @author Fersya
+ */
+public class EmployeeInfo {
+    private String employeeId;
+	private String firstName;
+	private String lastName;
+	private String idNumber;
+	private String address;
+        
+        public String getIdNumber(){
+            return idNumber;
+        }
+
+}
+
