+++ src/lib/Gender.java
@@ -1,0 +1,16 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package lib;
+
+/**
+ *
+ * @author Fersya
+ */
+public class Gender {
+    private enum gender{
+            Male,
+            Female
+        }
+}
